fix: empty cells, fix: value
Empty cells returns empty object therefore !emptyCell is not true. (The object exists with empty array inside)
THe fix for that is to check its length.
Value was defined in several places as val insted of value and in addition, because value and player can be 'A tmp' or 'A' a condition  that checked for 'A' wasn't corect. the fix was to add all values and players a substr(0,1)