Fix BUG: Victory for O is not recognize
The bug:
The game state function check for win situation and when X wins it find it and there will be a Modl for enouncing that.
When O wins, game state doen't recognise it.
The problem:
AI() called togelPlayer and past an argument of position. but togglePlayer() has empty argument and insted a global parmaeter _lasPos is used. Therefore the last position cell was used insted the corredt one.
The solution:
passing the corect ai selected move to global lasPos variable,